refactor -> add addUserBoxToList, add UserBox with removed content in UserListBox
